+++ src/modules/settings.ts
@@ -162,6 +162,9 @@ class RunnerSection extends SettingsSection {
     get timeAddition(): number {
         return this.get('timeAddition') as number;
     }
+    get stdoutThreshold(): number {
+        return this.get('stdoutThreshold') as number;
+    }
     get stderrThreshold(): number {
         return this.get('stderrThreshold') as number;
     }